+++ boundlexx/api/tasks.py
@@ -25,8 +25,9 @@ MAX_SINGLE_PURGE = 50
 WARM_CACHE_TASK = "boundlexx.api.tasks.warm_cache"
 PURGE_DJ_CACHE = "boundlexx.api.tasks.purge_django_cache"
 WARM_CACHE_PATHS = [
+    "/api/v1/schema/",
     "/api/v1/schema/?format=openapi-json",
-    "/api/v1/worlds/dump/",
+    "/api/v2/schema/",
     "/api/v2/schema/?format=openapi-json",
 ]